+++ dms-enterprise-20181101/ChangeLog.md
@@ -1,3 +1,8 @@
+2026-04-16 Version: 2.4.2
+- Update API ChatWithDesensitizeSSE: add request parameters ToolChoice.
+- Update API ChatWithDesensitizeSSE: add request parameters Tools.
+
+
 2026-04-08 Version: 2.4.1
 - Update API ListDBTaskSQLJobDetail: add response parameters Body.DBTaskSQLJobDetailList.$.SqlId.